aider: Add the `ttydb.svg` logo next to the Vue logo and link it to `talktoyourdatabase.com`.
# Context:
USER: add the ./assets/ttydb.svg logo next to the vue logo and link it to talktoyourdatabase.com
ASSISTANT: To add the `ttydb.svg` logo next to the Vue logo and link it to `https://talktoyourdatabase.com`, you will need to modify the `App.vue` file. Please ensure that the `ttydb.svg` file is placed in the `./assets/` directory.

Here are the changes you need to make to the `App.vue` file:

```vue
fe-clients/vue-ts/src/App.vue
<<<<<<< ORIGINAL
    <a href="https://vuejs.org/" target="_blank">
      <img src="./assets/vue.svg" class="logo vue" alt="Vue logo" />
    </a>
=======
    <a href="https://vuejs.org/" target="_blank">
      <img src="./assets/vue.svg" class="logo vue" alt="Vue logo" />
    </a>
    <a href="https://talktoyourdatabase.com" target="_blank">
      <img src="./assets/ttydb.svg" class="logo ttydb" alt="TTYDB logo" />
    </a>
>>>>>>> UPDATED
```

And add the following CSS to style the new logo:

```vue
fe-clients/vue-ts/src/App.vue
<<<<<<< ORIGINAL
/* Existing styles */
.logo {
  height: 6em;
  padding: 1.5em;
  will-change: filter;
  transition: filter 300ms;
}
=======
/* Existing styles */
.logo {
  height: 6em;
  padding: 1.5em;
  will-change: filter;
  transition: filter 300ms;
}

/* New styles for the TTYDB logo */
.logo.ttydb {
  height: 6em; /* Adjust the size as needed */
}
>>>>>>> UPDATED
```

Make sure the `ttydb.svg` file is present in the `./assets/` directory. If it's not there, you'll need to add it. If you need help with that or any other changes, let me know!